+++ src/usdb_syncer/resource_dl.py
@@ -1,10 +1,12 @@
 """Functions for downloading and processing media."""
 
+import io
 import os
 import traceback
+from dataclasses import dataclass
 from enum import Enum
 from pathlib import Path
-from typing import Union, assert_never
+from typing import assert_never
 
 import filetype
 import requests
@@ -18,6 +20,8 @@ from mutagen.oggvorbis import OggVorbis
 from PIL import Image, ImageEnhance, ImageOps
 from PIL.Image import Resampling
 
+from usdb_syncer import utils
+from usdb_syncer.constants import YtErrorMsg
 from usdb_syncer.download_options import AudioOptions, VideoOptions
 from usdb_syncer.logger import Log, song_logger
 from usdb_syncer.meta_tags import ImageMetaTags
@@ -42,7 +46,25 @@ DEFAULT_TARGET_LEVEL_R128 = -23.0  # dB
 DEFAULT_TARGET_LOUDNESS_RANGE = 7.0
 DEFAULT_TRUE_PEAK = -2.0
 
-YtdlOptions = dict[str, Union[str, bool, tuple, list, int]]
+YtdlOptions = dict[str, str | bool | tuple | list | int]
+
+
+class ResourceDLError(Enum):
+    """Errors that can occur when downloading a resource."""
+
+    RESOURCE_INVALID = "resource invalid"
+    RESOURCE_UNSUPPORTED = "resource unsupported"
+    RESOURCE_GEO_RESTRICTED = "resource geo-restricted"
+    RESOURCE_UNAVAILABLE = "resource unavailable"
+    RESOURCE_DL_FAILED = "resource download failed"
+
+
+@dataclass
+class ResourceDLResult:
+    """The result of a download operation."""
+
+    extension: str | None = None
+    error: ResourceDLError | None = None
 
 
 class ImageKind(Enum):
@@ -63,7 +85,7 @@ class ImageKind(Enum):
 
 def download_audio(
     resource: str, options: AudioOptions, browser: Browser, path_stem: Path, logger: Log
-) -> str | None:
+) -> ResourceDLResult:
     """Download audio from resource to path and process it according to options.
 
     Parameters:
@@ -73,7 +95,7 @@ def download_audio(
         path_stem: the target on the file system *without* an extension
 
     Returns:
-        the extension of the successfully downloaded file or None
+        DownloadResult with the extension of the downloaded file if successful
     """
     ydl_opts = _ytdl_options(
         options.ytdl_format(), browser, path_stem, options.rate_limit
@@ -89,17 +111,23 @@ def download_audio(
         }
         ydl_opts["postprocessors"] = [postprocessor]
 
-    if not (filename := _download_resource(ydl_opts, resource, logger)):
-        return None
-    if options.normalization in {
-        AudioNormalization.DISABLE,
-        AudioNormalization.REPLAYGAIN,
-    }:
-        filename = str(Path(filename).with_suffix(f".{options.format.value}"))
-    if options.normalization is not AudioNormalization.DISABLE:
-        _normalize(options, path_stem, filename, logger)
+    dl_result = _download_resource(resource, ydl_opts, logger)
+    if not dl_result.extension:
+        return dl_result
+    match options.normalization:
+        case AudioNormalization.DISABLE:
+            pass
+        case AudioNormalization.REPLAYGAIN:
+            filename = f"{path_stem}.{options.format.value}"
+            _normalize(options, path_stem, filename, logger)
+        case AudioNormalization.NORMALIZE:
+            filename = f"{path_stem}.{dl_result.extension}"
+            _normalize(options, path_stem, filename, logger)
+        case _ as unreachable:
+            assert_never(unreachable)
 
-    return options.format.value
+    dl_result.extension = options.format.value
+    return dl_result
 
 
 def _normalize(
@@ -220,7 +248,7 @@ def _write_replaygain_tags_opus(audio_file: str, track_gain: float) -> None:
 
 def download_video(
     resource: str, options: VideoOptions, browser: Browser, path_stem: Path, logger: Log
-) -> str | None:
+) -> ResourceDLResult:
     """Download video from resource to path and process it according to options.
 
     Parameters:
@@ -230,14 +258,12 @@ def download_video(
         path_stem: the target on the file system *without* an extension
 
     Returns:
-        the extension of the successfully downloaded file or None
+        DownloadResult with the extension of the downloaded file if successful
     """
     ydl_opts = _ytdl_options(
         options.ytdl_format(), browser, path_stem, options.rate_limit
     )
-    if filename := _download_resource(ydl_opts, resource, logger):
-        return os.path.splitext(filename)[1][1:]
-    return None
+    return _download_resource(resource, ydl_opts, logger)
 
 
 def _ytdl_options(
@@ -259,29 +285,66 @@ def _ytdl_options(
     return options
 
 
-def _download_resource(options: YtdlOptions, resource: str, logger: Log) -> str | None:
+def _download_resource(
+    resource: str, options: YtdlOptions, logger: Log
+) -> ResourceDLResult:
     if (url := video_url_from_resource(resource)) is None:
-        logger.debug(f"invalid audio/video resource: {resource}")
-        return None
+        return ResourceDLResult(error=ResourceDLError.RESOURCE_INVALID)
 
     options_without_cookies = options.copy()
-    options_without_cookies.pop("cookiesfrombrowser")
+    options_without_cookies.pop("cookiesfrombrowser", None)
+
     with yt_dlp.YoutubeDL(options_without_cookies) as ydl:
         try:
-            return ydl.prepare_filename(ydl.extract_info(url))
+            filename = ydl.prepare_filename(ydl.extract_info(url))
+            ext = os.path.splitext(filename)[1][1:]
+            return ResourceDLResult(extension=ext)
+        except yt_dlp.utils.UnsupportedError:
+            return ResourceDLResult(error=ResourceDLError.RESOURCE_UNSUPPORTED)
         except yt_dlp.utils.YoutubeDLError as e:
-            logger.debug(f"error downloading video url: {url}")
-            # Check if the error is due to age restriction
-            if "confirm your age" in str(e).lower():
-                logger.debug("Age-restricted resource. Retrying with cookies...")
-                try:
-                    with yt_dlp.YoutubeDL(options) as ydl:
-                        return ydl.prepare_filename(ydl.extract_info(url))
-                except yt_dlp.utils.YoutubeDLError as retry_error:
-                    logger.error(f"Retry failed: {retry_error}")
-                    return None
-            else:
-                return None
+            error_message = utils.remove_ansi_codes(str(e))
+            logger.debug(f"Failed to download '{url}': {error_message}")
+            if YtErrorMsg.YT_AGE_RESTRICTED in error_message:
+                dl_result = _retry_with_cookies(url, options, logger)
+                return ResourceDLResult(extension=dl_result.extension)
+            if YtErrorMsg.YT_GEO_RESTRICTED in error_message:
+                _handle_geo_restriction(url, resource, logger)
+                return ResourceDLResult(error=ResourceDLError.RESOURCE_GEO_RESTRICTED)
+            if YtErrorMsg.YT_UNAVAILABLE in error_message:
+                _handle_unavailable(url, logger)
+                return ResourceDLResult(error=ResourceDLError.RESOURCE_UNAVAILABLE)
+            raise
+
+
+def _retry_with_cookies(
+    url: str, options: YtdlOptions, logger: Log
+) -> ResourceDLResult:
+    logger.warning("Age-restricted resource. Retrying with cookies...")
+    with yt_dlp.YoutubeDL(options) as ydl:
+        try:
+            filename = ydl.prepare_filename(ydl.extract_info(url))
+            ext = os.path.splitext(filename)[1][1:]
+            return ResourceDLResult(extension=ext)
+        except yt_dlp.utils.YoutubeDLError as re:
+            logger.error(f"Retry failed: {utils.remove_ansi_codes(str(re))}")
+            raise
+
+
+def _handle_geo_restriction(url: str, resource: str, logger: Log) -> None:
+    logger.warning("Geo-restricted resource. You can retry after connecting to a VPN.")
+    if "youtube" in url and (
+        allowed_countries := utils.get_allowed_countries(resource)
+    ):
+        logger.info(
+            "Countries where the resource is available: " + ", ".join(allowed_countries)
+        )
+
+
+def _handle_unavailable(url: str, logger: Log) -> None:
+    logger.warning(
+        f"Resource '{url}' is no longer available. Please support the community, find a "
+        "suitable replacement resource and comment it on USDB."
+    )
 
 
 def download_image(url: str, logger: Log) -> bytes | None:
@@ -300,8 +363,15 @@ def download_image(url: str, logger: Log) -> bytes | None:
             "down or your internet connection is currently unavailable."
         )
         return None
-    if reply.status_code in range(100, 399):
-        # 1xx informational response, 2xx success, 3xx redirection
+    if reply.status_code in range(100, 299):
+        # 1xx informational response, 2xx success
+        return reply.content
+    if reply.status_code in range(300, 399):
+        # 3xx redirection
+        logger.debug(
+            f"'{url}' redirects to '{reply.headers["Location"]}'. "
+            "Please adapt metatags."
+        )
         return reply.content
     if reply.status_code in range(400, 499):
         logger.error(
@@ -330,37 +400,65 @@ def download_and_process_image(
         return None
 
     if not filetype.is_image(img_bytes):
-        logger.error(f"#{str(kind).upper()}: file at {url} is no image")
+        logger.error(f"#{str(kind).upper()}: file at {url} is not an image")
         return None
 
     path = target_stem.with_name(f"{target_stem.name} [{kind.value}].jpg")
-    with path.open("wb") as file:
-        file.write(img_bytes)
+    with Image.open(io.BytesIO(img_bytes)).convert("RGB") as img:
+        img.save(path)
 
     if process:
-        _process_image(meta_tags, max_width, path)
+        _process_image(meta_tags, kind, max_width, path)
     return path
 
 
+def _rotate(image: Image.Image, meta_tags: ImageMetaTags) -> Image.Image:
+    if rotate := meta_tags.rotate:
+        return image.rotate(rotate, resample=Resampling.BICUBIC, expand=True)
+    return image
+
+
+def _crop(image: Image.Image, meta_tags: ImageMetaTags) -> Image.Image:
+    if crop := meta_tags.crop:
+        return image.crop((crop.left, crop.upper, crop.right, crop.lower))
+    return image
+
+
+def _resize(image: Image.Image, meta_tags: ImageMetaTags) -> Image.Image:
+    if resize := meta_tags.resize:
+        return image.resize((resize.width, resize.height), resample=Resampling.LANCZOS)
+    return image
+
+
+def _adjust_contrast(image: Image.Image, meta_tags: ImageMetaTags) -> Image.Image:
+    if meta_tags.contrast == "auto":
+        return ImageOps.autocontrast(image, cutoff=5)
+    if meta_tags.contrast:
+        return ImageEnhance.Contrast(image).enhance(meta_tags.contrast)
+    return image
+
+
 def _process_image(
-    meta_tags: ImageMetaTags | None, max_width: CoverMaxSize | None, path: Path
+    meta_tags: ImageMetaTags | None,
+    kind: ImageKind,
+    max_width: CoverMaxSize | None,
+    path: Path,
 ) -> None:
     processed = False
     with Image.open(path).convert("RGB") as image:
         if meta_tags and meta_tags.image_processing():
             processed = True
-            if rotate := meta_tags.rotate:
-                image = image.rotate(rotate, resample=Resampling.BICUBIC, expand=True)
-            if crop := meta_tags.crop:
-                image = image.crop((crop.left, crop.upper, crop.right, crop.lower))
-            if resize := meta_tags.resize:
-                image = image.resize(
-                    (resize.width, resize.height), resample=Resampling.LANCZOS
-                )
-            if meta_tags.contrast == "auto":
-                image = ImageOps.autocontrast(image, cutoff=5)
-            elif meta_tags.contrast:
-                image = ImageEnhance.Contrast(image).enhance(meta_tags.contrast)
+            match kind:
+                case ImageKind.COVER:
+                    operations = [_rotate, _crop, _resize, _adjust_contrast]
+                case ImageKind.BACKGROUND:
+                    operations = [_resize, _crop]
+                case _ as unreachable:
+                    assert_never(unreachable)
+
+            for operation in operations:
+                image = operation(image, meta_tags)
+
         if (
             max_width
             and max_width != CoverMaxSize.DISABLE